+++ packages/terra-core-docs/src/terra-dev-site/doc/html-table/Table.1.doc.mdx
@@ -1,11 +1,11 @@
 import { Badge } from 'terra-html-table/package.json?dev-site-package';
 
-import DefaultTable from '../example/DefaultTable?dev-site-example';
-import NoStripedTable from '../example/NoStripedTable?dev-site-example';
-import CompactPaddingTable from '../example/CompactPaddingTable?dev-site-example';
-import StandardPaddingTable from '../example/StandardPaddingTable?dev-site-example';
-import TableWithLongContent from '../example/TableWithLongContent?dev-site-example';
-import TableWithCustomCells from '../example/TableWithCustomCells?dev-site-example';
+import DefaultTable from './example/DefaultTable?dev-site-example';
+import NoStripedTable from './example/NoStripedTable?dev-site-example';
+import CompactPaddingTable from './example/CompactPaddingTable?dev-site-example';
+import StandardPaddingTable from './example/StandardPaddingTable?dev-site-example';
+import TableWithLongContent from './example/TableWithLongContent?dev-site-example';
+import TableWithCustomCells from './example/TableWithCustomCells?dev-site-example';
 
 import TablePropsTable from 'terra-html-table/src/Table?dev-site-props-table';
 import TableHeaderPropsTable from 'terra-html-table/src/TableHeader?dev-site-props-table';
