+++ packages/terra-core-docs/src/terra-dev-site/doc/html-table/example/StandardPaddingTable.jsx
