+++ packages/terra-core-docs/src/terra-dev-site/test/html-table/CompactPaddingTable.test.jsx
