+++ packages/terra-core-docs/src/terra-dev-site/test/html-table/NoPaddingTable.test.jsx
@@ -5,7 +5,7 @@ import Table, {
   Body,
   Cell,
   Row,
-} from '../../../index';
+} from 'terra-html-table';
 
 const NoPaddingTable = () => (
   <Table paddingStyle="none">
