+++ packages/terra-core-docs/src/terra-dev-site/test/html-table/StandardPaddingTable.test.jsx
