+++ packages/terra-core-docs/src/terra-dev-site/test/html-table/StripedTable.test.jsx
@@ -5,23 +5,33 @@ import Table, {
   HeaderCell,
   Row,
   Cell,
-} from '../../../index';
+} from 'terra-html-table';
 
-const SingleRowTable = () => (
+const StripedTable = () => (
   <Table>
     <Header>
       <HeaderCell key="NAME">Name</HeaderCell>
       <HeaderCell key="ADDRESS">Address</HeaderCell>
       <HeaderCell key="PHONE_NUMBER">Phone Number</HeaderCell>
     </Header>
-    <Body className="TableRows">
+    <Body>
       <Row key="PERSON_0">
         <Cell key="NAME">John Smith</Cell>
         <Cell key="ADDRESS">123 Adams Drive</Cell>
+        <Cell key="PHONE_NUMBER">111-222-3333</Cell>
+      </Row>
+      <Row key="PERSON_1">
+        <Cell key="NAME">Jane Smith</Cell>
+        <Cell key="ADDRESS">321 Drive Street</Cell>
+        <Cell key="PHONE_NUMBER">111-222-3333</Cell>
+      </Row>
+      <Row key="PERSON_2">
+        <Cell key="NAME">Dave Smith</Cell>
+        <Cell key="ADDRESS">213 Raymond Road</Cell>
         <Cell key="PHONE_NUMBER">111-222-3333</Cell>
       </Row>
     </Body>
   </Table>
 );
 
-export default SingleRowTable;
+export default StripedTable;
