+++ packages/terra-core-docs/src/terra-dev-site/test/html-table/TableNoRows.test.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import Table, {
+  Body,
+  Header,
+  HeaderCell,
+} from 'terra-html-table';
+
+const SingleRowTable = () => (
+  <Table>
+    <Header>
+      <HeaderCell key="NAME">Name</HeaderCell>
+      <HeaderCell key="ADDRESS">Address</HeaderCell>
+      <HeaderCell key="PHONE_NUMBER">Phone Number</HeaderCell>
+    </Header>
+    <Body className="TableRows" />
+  </Table>
+);
+
+export default SingleRowTable;
